+++ illustrations_achievements.php
@@ -24,20 +24,58 @@
 
     <?php include_once 'partials/headers.php' ?>
 
+    <div id="popup-container" class="invisible">
+        <img id="popup-img" class="fullscreen-child popup-img" src="">
+        <div id="top-right-x">X</div>
+    </div>
+
     <div class="main_content">
 
-        <div class="add_padding card-nomargin">
-
+        <div class="add_padding card-firstdiv">
+            
             <h1 id="illu_achiev_id">Illustrations</h1>
+            
+            <div id="button_container_horizontal" class="flex-container" style="margin-bottom: 20px;"> <!-- les boutons jaunes -->
+            
+            <div class="flex-panel2"></div>
+            <div id="illustrations-button" class="flex-panel2 button yellow_button" style="margin: 10px" onclick="refresh_lines(0)">Illustrations</div>
+            <div id="achievements-button" class="flex-panel2 button yellow_button" style="margin: 10px" onclick="refresh_lines(1)">Achievements</div>
+            <div class="flex-panel2"></div>
 
-            <div id="button_container_horizontal"> <!-- les boutons jaunes -->
-                
             </div>
 
-            <div id="lines_container">
-                
+            <div id="illustrations">
+                <div class="x-card-purple">
+                    <div class="x-card-purple-header">
+                        <h4>Chapter 1</h4>
+                    </div>
+
+                    <div class="x-card-content flex-container">
+                        <div class="card-left-content"><img class="x-card-img-purple" src="images/game_images/illustrations/chapter1/illustration1.jpg" onclick="showPopup(this.src)"></img></div>
+                        <div class="card-middle-content"><img class="x-card-img-purple" src="images/game_images/illustrations/chapter1/illustration2.jpg" onclick="showPopup(this.src)"></img></div>
+                        <div class="card-middle-content"><img class="x-card-img-purple" src="images/game_images/illustrations/chapter1/illustration3.jpg" onclick="showPopup(this.src)"></img></div>
+                        <div class="card-right-content"><img class="x-card-img-purple" src="images/game_images/illustrations/chapter1/illustration4.jpg" onclick="showPopup(this.src)"></img></div>
+                    </div>
+                </div>		
+
+                <div class="x-card-pink">
+                    <div class="x-card-pink-header">
+                        <h4>Chapter 2</h4>
+                    </div>
+
+                    <div class="x-card-content flex-container">
+                        <div class="card-left-content"><img class="x-card-img-pink" src="images/game_images/illustrations/chapter2/illustration1.jpg" onclick="showPopup(this.src)"></img></div>
+                        <div class="card-middle-content"><img class="x-card-img-pink" src="images/game_images/illustrations/chapter2/illustration2.jpg" onclick="showPopup(this.src)"></img></div>
+                        <div class="card-middle-content"></div>
+                        <div class="card-right-content"></div>
+                    </div>
+                </div>		
             </div>
             
+            <div id="achievements">
+
+            </div>
+                
         </div>
     </div>
 
@@ -45,6 +83,9 @@
 
 </body>
 
+<script src = "js/jquery_3.2.1.js"></script>
+<script src = "js/variables.js"></script>
+<script src = "js/library.js"></script>
 <script src="js/illustrations_achievements.js"></script>
 
 </html>