+++ relationships.php
@@ -21,9 +21,8 @@
 
     <?php include_once 'partials/headers.php' ?>
 
-
     <div id="popup-container" class="invisible">
-        <img id="school_file_img_container" class="fullscreen-child" src="_new_images_folder/characters/school_files/alistair.png">
+        <img id="popup-img" class="fullscreen-child" src="_new_images_folder/characters/school_files/alistair.png">
         <div id="top-right-x">X</div>
     </div>